Move config functions
