+++ .env.dist
@@ -1,3 +1,4 @@
+CONFIG_PATH=/config
 PORT=3000
-QBIT_BASE=http://localhost:8080
-VUETORRENT_PATH=./data/vuetorrent
+QBIT_BASE=http://host.docker.internal:8080
+VUETORRENT_PATH=/vuetorrent
